agregado de test de figuras y se saca figurasSeSolapan. Falta test grFIG

+++ Composite/src/edu/unlam/paradigmas/composite/Cuadrado.java
@@ -3,9 +3,10 @@ package edu.unlam.paradigmas.composite;
 public class Cuadrado extends Figura {
 	private double lado;
 
-	public Cuadrado(int lado, String nombre, String color,Punto ubicacion) throws Exception {
+	public Cuadrado(double lado, String nombre, String color, Punto ubicacion) throws Exception {
 		super(nombre, color, ubicacion);
-		this.lado = validarAtributosFigura(lado,"Por favor, revise que el lado del cuadrado " + this.nombre + " sea mayor o menor que cero centimetros");
+		this.lado = validarAtributosFigura(lado, "Por favor, revise que el lado del cuadrado " + this.nombre
+				+ " sea mayor o menor que cero centimetros");
 	}
 
 	@Override
